Writer: Fix sidebar table btns position shift
Remove unnecessary rules targetting minimalrowheight and
minimalcolumnwidth. These rules were
1. Not being apply always (sometimes jsbuilder was
setting them with different id, e.g.: with suffix 1)
2. Even when apply this rules were forcing these two btns
to go up, looking like they are mere icon for the spinner.
This is wrong. They are buttons that trigger the set to the
minimal sizes of that particular row or columns