calc: pinch-zoom: fix hard edge when zoom-panning
Previously we clamped the pinch so it could not go outside the bounds of
the initial zoom pane (presumably this was intended to avoid you going
outside the map with your pinch).

Unfortunately, this is the initial bounds, causing a nasty edge when you
panned around too much, which is particularly noticable with the
improved zoom-pan ergonomics.

I decided to remove this altogether rather than scaling it correctly, as
I don't believe that a check to make sure you're not going outside the
map feels particularly natural when pinch-zooming around the document,
particularly as only the center of your pinch is checked.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Ib98ca91e82bbe7f2bb9986595fc5f7928e2206c9