postmessage demo: insert graphic test
Add button in the post message tester to verify if
inserting an image works properly. I used resource
from our SDK so it is external domain, if not allowed
it shows an error. We need to add domain to oxoolwsd.xml
in the lok_allow section to allow importing the image.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: Iaf818dcda9246c5006fd678f856c8b90dab675fb